server side validation to delete spots from admin page

+++ pages/api/admin/[spot].js
@@ -8,13 +8,17 @@ export default async (req, res) => {
     switch (method) {
         case "POST":
             try {
-                Spot.findByIdAndDelete(req.query.spot, (err) => {
-                    if (err) {
-                        return next(err);
-                    }
-                    res.redirect("/overview");
-                });
-                res.status(201).json({ success: true });
+                if (req.query.user === process.env.NEXT_PUBLIC_ADMIN_ID) {
+                    Spot.findByIdAndDelete(req.query.spot, (err) => {
+                        if (err) {
+                            return next(err);
+                        }
+                    });
+                    res.status(201).json({ success: true });
+                } else {
+                    res.status(400).json({ success: false });
+                }
+
             } catch (error) {
                 console.log(error)
                 res.status(400).json({ success: false });
